requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.23.2
pydantic==1.10.5
mysql-connector-python==8.0.33
python-dotenv==1.0.0
python-multipart